lircd: fix things up so decode behaves correctly
There were a few things I changed to __u32 earlier that should have been
left as unsigned long, and the old #ifdef __GLIBC__ hack appears to be
no longer necessary with ir_code as a __u64 (could be glibc got fixed,
not sure, but it works, so lets drop the cruft). Also converts a few
things that are type lirc_t that store return values from functions that
returned unsigned longs (now __u32's, mostly) over to __u32 types.

With this, lirc encode and decode behave 100% identical with 0.8.7
again, at least in testing on a 64-bit system...

+++ daemons/ir_remote.h
@@ -304,10 +304,10 @@ static inline lirc_t lower_limit(struct ir_remote *remote, lirc_t val)
 }
 
 /* only works if last <= current */
-static inline __u32 time_elapsed(struct timeval *last,
-				 struct timeval *current)
+static inline unsigned long time_elapsed(struct timeval *last,
+					 struct timeval *current)
 {
-	__u32 secs,diff;
+	unsigned long secs,diff;
 	
 	secs=current->tv_sec-last->tv_sec;
